routes/
The routes folder is where I define what happens when someone visits different pages on my portfolio. For example, if someone clicks on the "About Me" link, the route file will control which content gets displayed on that page. It’s the part of the project that makes my portfolio interactive and dynamic.