package.json
The package.json file is like a blueprint for my project. It lists all the dependencies (like Express, Bootstrap, etc.) that my portfolio needs to run. It also includes scripts for starting the server and other tasks. It’s a central file that makes managing my project’s setup easier.